Move gov code to its own package
